Rename and tidy up DeletedCommentsCounter
